Raise an exception when reconnection fails
This hopefully prevents the situation in which
an unusable (e.g. closed) connection is returned to the pool,
causing the next client in queue fail to execute its query.

The expectation is, by propagating the `DBConnection.ConnectionError`
we'll make it caught by DBConnection and the broken connection will
be removed from the pool.